Refactor: strictly identified and left only methods that communicate directly with DB.
